useGlobalState
1. removed reducer logic from useRowActions
2. created useGLobalState hook
3. reducer-actions index exports necessary things
4. passed to row
