+++ src/reducers-actions/index.ts
@@ -1,0 +1,6 @@
+import { TokenLocalState } from './tokenRow'
+
+export * from './tokenRow'
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface GlobalState extends TokenLocalState {}
